replace navigation callbacks
